requirements/prod.txt
aiohttp==3.5.4
alembic==1.0.6
appnope==0.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
gofer-grader==1.0.3
google-cloud-secret-manager==2.7.2
idna==2.8
idna-ssl==1.1.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
Jinja2==2.10
jupyterhub==0.9.4
lxml==4.6.3
Mako==1.0.7
MarkupSafe==1.1.0
multidict==4.5.2
oauthlib==3.0.0
otter-grader==3.1.4
pamela==0.3.0
pandas==1.3.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pygments==2.10.0
python-dateutil==2.7.5
python-dotenv==0.19.1
python-editor==1.0.3
python-oauth2==1.1.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.16
tornado==5.1.1
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.24.1
wcwidth==0.1.7
yarl==1.3.0